Renames OAuthRequests to OAuthSession
Also make it a reference type, which makes more sense.

There was some though to this being all immutable values, but it made a lot of code less-obvious. It's still effectively immutable as a class, just in a more obvious way.

Also cleans up naming to be more obvious throughout.